Create util folder and move related class into it.
