+++ src/ViewModels/ChampionshipResultConvertViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using RaceResultConverter.Enum;
+using RaceResultConverter.Utils;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.IO;
