+++ src/ViewModels/SingleRaceConvertViewModel.cs
@@ -5,6 +5,7 @@ using RaceResultConverter.Conversion;
 using RaceResultConverter.DTO.Zon;
 using RaceResultConverter.DTO.ZRound;
 using RaceResultConverter.Enum;
+using RaceResultConverter.Utils;
 using OpenFileDialog = Microsoft.Win32.OpenFileDialog;
 using SaveFileDialog = Microsoft.Win32.SaveFileDialog;
